+++ src/structs.rs
@@ -10,7 +10,7 @@ pub struct LogData {
 #[serde(rename_all = "snake_case")]
 pub enum TimerResponse {
     StartUpdate {
-        esp_id: u128,
+        esp_id: u64,
         version: String,
         build_time: u128, // NOT USED
         size: i64,
@@ -20,34 +20,34 @@ pub enum TimerResponse {
         offset: i64,
         competitor_id: u128,
         judge_id: u128,
-        esp_id: u128,
+        esp_id: u64,
         timestamp: u128,
         session_id: String, // UUID
         delegate: bool,
     },
     SolveConfirm {
-        esp_id: u128,
+        esp_id: u64,
         competitor_id: u128,
         session_id: String,
     },
     ApiError {
-        esp_id: u128,
+        esp_id: u64,
         error: String,
         should_reset_time: bool,
     },
     CardInfoRequest {
         card_id: u128,
-        esp_id: u128,
+        esp_id: u64,
     },
     CardInfoResponse {
         card_id: u128,
-        esp_id: u128,
+        esp_id: u64,
         display: String,
         country_iso2: String,
         can_compete: bool,
     },
     Logs {
-        esp_id: u128,
+        esp_id: u64,
         logs: Vec<LogData>,
     },
 }
